Complete first draft of the game
The program is divided in 3 main classes, Board, Player and Game.
PlayerA and PlayerB are subclasses of Player and are used to keep track
of who's turn ut is.
Board class simply displays the game board.
Game class contains the logic of the game, the method play calls the
method turn 9 times. Each turn call gets the turn player input and calls
the set_mark_position, it then displays the updated board.